% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{format_Matrix}
\alias{format_Matrix}
\title{Imbue Matrix With Property Classes}
\usage{
format_Matrix(x, sparse = TRUE, symmetric = FALSE,
  triangular = FALSE, diagonal = FALSE)
}
\arguments{
\item{x}{a matrix with (or without) zero values}

\item{sparse}{logical; gives class sparseMatrix and is almost always true.}

\item{symmetric}{logical; gives class symmetricMatrix}

\item{triangular}{logical; gives class triangularMatrix}

\item{diagonal}{logical; gives class diagonalMatrix}
}
\description{
Gives a matrix classes such as sparseness and symmetry from the Matrix package. Note that
a matrix cannot be both symmetric and triangular unless it is diagonal.
}
