% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{qform}
\alias{qform}
\title{Efficient Computation of Quadratic Form}
\usage{
qform(x, A, U)
}
\arguments{
\item{x}{a vector}

\item{A}{a matrix. Only one of A and U needs to be provided.}

\item{U}{an upper triangular matrix. Note that plugging in chol(A) is slower.}
}
\description{
Evaluates the quadratic form x'Ax using optimized cross products.
An upper triangular matrix with U'U = A can also be given to dramatically increase speed.
}
