% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{sqrt.eigen}
\alias{sqrt.eigen}
\title{Square Root Matrix using Eigen Values/Vectors}
\usage{
\method{sqrt}{eigen}(x, symmetric = FALSE)
}
\arguments{
\item{x}{semi-positive definite matrix}

\item{symmetric}{logical; indicates to eigen if x is symmetric}
}
\description{
Compute a square root matrix using the eigenvalue/vector spectral decomposition.
}
